requirements.txt
django
pyserial
djangorestframework
django-cors-headers